requirements.txt
alembic==1.16.1
amqp==5.3.1
anyio==4.9.0
async-timeout==5.0.1
asyncpg==0.28.0
bcrypt==4.3.0
billiard==4.2.1
celery==5.3.0
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
dnspython==2.7.0
ecdsa==0.19.1
email_validator==2.2.0
fastapi==0.95.2
greenlet==3.2.2
h11==0.16.0
httptools==0.6.4
idna==3.10
kombu==5.5.3
Mako==1.3.10
MarkupSafe==3.0.2
passlib==1.7.4
prompt_toolkit==3.0.51
psycopg2-binary==2.9.10
pyasn1==0.6.1
pydantic==1.10.9
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-jose==3.3.0
PyYAML==6.0.2
redis==4.5.1
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.19
starlette==0.27.0
typing_extensions==4.13.2
tzdata==2025.2
uvicorn==0.22.0
uvloop==0.21.0
vine==5.1.0
watchfiles==1.0.5
wcwidth==0.2.13
websockets==15.0.1
